added primative text field, used erics background and davids fisherman as a test for pixel sizes
