+++ Hero'sJourney/src/island/Island.java
@@ -30,8 +30,11 @@ public class Island extends JPanel implements ActionListener, KeyListener, Mouse
 	Extra e1;
 	Ninja n;
 	
+	Text ninjaText;
+	
 	ArrayList<Music> bg = new ArrayList<Music>();
 	Font verdana = new Font("Verdana", Font.BOLD, 40);
+	Font verdanaSmall = new Font("Verdana", Font.BOLD, 20);
 	
 	public int x, y;
 	public int vx, vy;
@@ -77,7 +80,7 @@ public class Island extends JPanel implements ActionListener, KeyListener, Mouse
 		//call paint methods of objects
 		i.paint(g);
 		e1.paint(g);
-		n.paint(g);
+		//n.paint(g);
 		
 		//paint player last
 		p.paint(g);
@@ -100,7 +103,9 @@ public class Island extends JPanel implements ActionListener, KeyListener, Mouse
 		//shuffle
 		shuffleMusic();
 		
-		n.speak(g, p, verdana, width, height);
+		if (n.isIntersectN(p)) {
+			ninjaText.print(g, verdanaSmall, width, height/4);
+		}
 		
 		g.setColor(Color.red);
 		g.drawLine(midX - 25, 0, midX - 25, height);	//left
@@ -218,10 +223,15 @@ public class Island extends JPanel implements ActionListener, KeyListener, Mouse
 		width = f.getWidth();
 		height = f.getHeight();
 		
-		i = new IslandBackground("testIsland.png", width*2, height*2);
-		e1 = new Extra("stego.png", 850, 850, 50, 50);
+		i = new IslandBackground("BackgroundSandTEst.png", width*4, height*4);
+		e1 = new Extra("fisherman.png", 850, 900, 50, 50);
 		n = new Ninja("ninja.png", 1000, 600, 150, 150);
 		p = new Protagonist("bronc.png", midX - 25, midY - 25, 50, 50);
+		
+		ninjaText = new Text("and then he touched with his lips, \r\n" + 
+				"together we became. One Forever. \r\n" + 
+				"And when he took of his shirt \r\n" +
+				"I laughed fo he was an outie");
 		
 		bg.add(new Music("Gravity.wav", true, "Gravity by Brent Faiyaz"));
 		bg.add(new Music("Blessed.wav", true, "Blessed by Juls"));
